// admin/src/app/models/hotel-images.model.ts
export class Hotel_Images {
    Id: null;
    LocationID: number;
    CountryID: number;
    CountryLocationID: number;
    CityID: number;
    CityLocationID: number;
    HotelID: number;
    HotelLocationID: number;
    HotelImages: string;
    ImagesDescription: string;
    LanguageValue: string;
    CreatedOn: null;

}